refactor: move Router and related files to its own package
