+++ composer.json
@@ -60,6 +60,7 @@
         "tempest/log": "self.version",
         "tempest/mapper": "self.version",
         "tempest/reflection": "self.version",
+        "tempest/routing": "self.version",
         "tempest/support": "self.version",
         "tempest/validation": "self.version",
         "tempest/view": "self.version"
@@ -81,6 +82,7 @@
             "Tempest\\Log\\": "src/Tempest/Log/src",
             "Tempest\\Mapper\\": "src/Tempest/Mapper/src",
             "Tempest\\Reflection\\": "src/Tempest/Reflection/src",
+            "Tempest\\Routing\\": "src/Tempest/Routing/src",
             "Tempest\\Support\\": "src/Tempest/Support/src",
             "Tempest\\Validation\\": "src/Tempest/Validation/src",
             "Tempest\\View\\": "src/Tempest/View/src"
@@ -109,6 +111,8 @@
             "Tempest\\Log\\Tests\\": "src/Tempest/Log/tests",
             "Tempest\\Mapper\\Tests\\": "src/Tempest/Mapper/tests",
             "Tempest\\Reflection\\Tests\\": "src/Tempest/Reflection/tests",
+            "Tempest\\Routing\\": "src/Tempest/Routing/tests",
+            "Tempest\\Routing\\Tests\\": "src/Tempest/Routing/tests",
             "Tempest\\Support\\Tests\\": "src/Tempest/Support/tests",
             "Tempest\\Validation\\Tests\\": "src/Tempest/Validation/tests",
             "Tests\\Tempest\\": "tests/"
