+++ src/Tempest/Http/src/Static/StaticGenerateCommand.php
@@ -12,9 +12,9 @@ use Tempest\Core\Kernel;
 use Tempest\Http\DataProvider;
 use Tempest\Http\GenericRequest;
 use Tempest\Http\Method;
-use Tempest\Http\Router;
 use Tempest\Http\Status;
 use function Tempest\path;
+use Tempest\Routing\Router;
 use function Tempest\uri;
 use Tempest\View\View;
 use Tempest\View\ViewRenderer;
