+++ src/Tempest/Routing/.gitattributes
@@ -1,0 +1,10 @@
+# Exclude build/test files from the release
+.github/                  export-ignore
+tests/                    export-ignore
+.gitattributes            export-ignore
+.gitignore                export-ignore
+phpunit.xml               export-ignore
+README.md                 export-ignore
+
+# Configure diff output for .php and .phar files.
+*.php diff=php
