+++ src/Tempest/Routing/composer.json
@@ -1,0 +1,30 @@
+{
+    "name": "tempest/routing",
+    "license": "MIT",
+    "minimum-stability": "dev",
+    "require": {
+        "php": "^8.3",
+        "tempest/core": "self.version",
+        "tempest/clock": "self.version",
+        "tempest/console": "self.version",
+        "tempest/view": "self.version",
+        "tempest/mapper": "self.version",
+        "tempest/container": "self.version",
+        "tempest/highlight": "^2.0",
+        "laminas/laminas-diactoros": "^3.3",
+        "psr/http-factory": "^1.0",
+        "psr/http-message": "^1.0|^2.0",
+        "ramsey/uuid": "^4.7",
+        "symfony/var-exporter": "^7.1"
+    },
+    "autoload": {
+        "psr-4": {
+            "Tempest\\Routing\\": "src"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Tempest\\Routing\\Tests\\": "tests"
+        }
+    }
+}
